refactor(WebAPI): Cleanup op new dash-mpd-cli version wrapper
